+++ TaskApp/Controllers/MissionApiController.cs
@@ -21,14 +21,13 @@ namespace TaskApp.Controllers
 		private readonly IMissionService _missionService;
 		private readonly IOperationService _operationService;
 
-
-
 		public MissionApiController(IUserService userService, IMissionService missionService, IOperationService operationService)
 		{
 			_userService = userService;
 			_missionService = missionService;
 			_operationService = operationService;
 		}
+
 		[HttpGet]
 		[Route(nameof(GetMyMissions))]
 		public ActionResult<ApiResponse<List<MissionModel>>> GetMyMissions()
@@ -36,6 +35,10 @@ namespace TaskApp.Controllers
 			try
 			{
 				var user = this._userService.GetOnlineUser(this.HttpContext);
+				if (user == null)
+				{
+					return Json(ApiResponse<List<MissionModel>>.WithError("Not authorized !"));
+				}
 				var missions = this._missionService.GetAllMyMissionsByUserId(user.Id);
 
 				var response = ApiResponse<List<MissionModel>>.WithSuccess(missions);
@@ -47,26 +50,15 @@ namespace TaskApp.Controllers
 				return Json(ApiResponse<List<MissionModel>>.WithError(exp.ToString()));
 			}
 		}
+
 		[HttpGet]
 		[Route(nameof(GetAllMissions))]
 		public ActionResult<ApiResponse<List<MissionModel>>> GetAllMissions()
 		{
 			try
 			{
-				
+				// DONE: döngü yerine join yap
 				var missions = this._missionService.GetAll().ToList();
-				var users = this._userService.GetAllUsers();
-				foreach(UserModel user in users)
-				{
-					for(int i=0; i<missions.Count;i++)
-					{
-						if(user.Id == missions[i].UserId)
-						{
-							missions[i].MissionUsername = user.Username;
-						}
-					}
-				}
-
 				var response = ApiResponse<List<MissionModel>>.WithSuccess(missions);
 
 				return Json(response);
@@ -76,13 +68,15 @@ namespace TaskApp.Controllers
 				return Json(ApiResponse<List<MissionModel>>.WithError(exp.ToString()));
 			}
 		}
+
+		// DONE: bu method missionId parametresi almalı, adı düzeltilmeli, javascriptteki kısım da düzeltilmeli
 		[HttpGet]
-		[Route(nameof(GetCurrentMissionOperations))]
-		public ActionResult<ApiResponse<List<OperationModel>>> GetCurrentMissionOperations()
+		[Route(nameof(GetOperationsByMissionId))]
+		public ActionResult<ApiResponse<List<OperationModel>>> GetOperationsByMissionId(int missionId)
 		{
 			try
 			{
-				var operations = this._operationService.GetAll();
+				var operations = this._operationService.GetOperationsByMissionId(missionId);
 				
 				var response = ApiResponse<List<OperationModel>>.WithSuccess(operations);
 
@@ -93,18 +87,23 @@ namespace TaskApp.Controllers
 				return Json(ApiResponse<List<OperationModel>>.WithError(exp.ToString()));
 			}
 		}
+
 		[HttpPost]
 		[Route(nameof(CreateMission))]
 		public ActionResult<ApiResponse<MissionModel>> CreateMission([FromBody]CreateMissionModel model)
 		{
 			try
 			{
-				if(model.MissionName == null || model.MissionName == "")
+				var user = this._userService.GetOnlineUser(this.HttpContext);
+				if (user == null)
+				{
+					return Json(ApiResponse<List<ForumPostModel>>.WithError("Not authorized !"));
+				}
+				if (model.MissionName == null || model.MissionName == "")
 				{
 					return Json(ApiResponse<UserModel>.WithError("Mission name is required !"));
 				}
 				MissionModel result = null;
-				var user = this._userService.GetOnlineUser(this.HttpContext);
 				var newMission = new Mission();
 				newMission.MissionName = model.MissionName;
 				newMission.UserId = user.Id;
@@ -112,7 +111,7 @@ namespace TaskApp.Controllers
 
 				this._missionService.AddNewMission(newMission);
 				result = this._missionService.GetById(newMission.Id);
-				result.MissionUsername = user.Username;
+
 				return Json(ApiResponse<MissionModel>.WithSuccess(result));
 			}
 			catch (Exception exp)
@@ -120,56 +119,75 @@ namespace TaskApp.Controllers
 				return Json(ApiResponse<MissionModel>.WithError(exp.ToString()));
 			}
 		}
+
 		[HttpPost]
 		[Route(nameof(AddOperation))]
 		public ActionResult<ApiResponse<OperationModel>> AddOperation([FromBody]CreateOperationModel model)
 		{
+			// DONE: sadece kendi mission ıma ekleyebilmeliyim
 			try
 			{
+				var user = this._userService.GetOnlineUser(this.HttpContext);
+				if (user == null)
+				{
+					return Json(ApiResponse<List<ForumPostModel>>.WithError("Not authorized !"));
+				}
 				if (model.OperationContent == null || model.OperationContent == "")
 				{
 					return Json(ApiResponse<UserModel>.WithError("Operation name is required !"));
 				}
-				var operations = this._operationService.GetAll();
-				int operationCount = 0;
-				foreach (OperationModel operation in operations)
+				var mission = _missionService.GetById(model.MissionId);
+				if(user.Id != mission.UserId)
 				{
-					if (operation.MissionId == Convert.ToInt32(model.MissionId))
-					{
-						operationCount++;
-					}
+					return Json(ApiResponse<List<ForumPostModel>>.WithError("Not authorized !"));
 				}
-				if (operationCount >= 9)
+				// DONE: burada döngü olmamalı, get ops by mission id olmalı
+				var operations = this._operationService.GetOperationsByMissionId(mission.Id);
+				if(operations.Count >= 9)
 				{
 					return Json(ApiResponse<UserModel>.WithError("You cant add operations more then 9 !"));
 				}
+				else
+				{
+					var newOperation = new Operation();
+					newOperation.OperationContent = model.OperationContent;
+					newOperation.MissionId = model.MissionId;
+					newOperation.OperationStatus = Enums.OperationStatus.NotDone;
 
-				var newOperation = new Operation();
-				newOperation.OperationContent = model.OperationContent;
-				newOperation.MissionId = Convert.ToInt32(model.MissionId);
-				newOperation.OperationStatus = 0;
-				
-				
-		
-				this._operationService.AddNewOperation(newOperation);
-				OperationModel result = null;
-				result = this._operationService.GetById(newOperation.Id);
-				return Json(ApiResponse<OperationModel>.WithSuccess(result));
+
+
+					this._operationService.AddNewOperation(newOperation);
+					OperationModel result = this._operationService.GetById(newOperation.Id);
+					return Json(ApiResponse<OperationModel>.WithSuccess(result));
+				}
+
 			}
 			catch (Exception exp)
 			{
 				return Json(ApiResponse<OperationModel>.WithError(exp.ToString()));
 			}
 		}
+
 		[HttpPost]
 		[Route(nameof(UpdateOperation))]
 		public ActionResult<ApiResponse<OperationModel>> UpdateOperation([FromBody]CreateOperationModel model)
 		{
+			
 			try
 			{
-				var operation = this._operationService.GetByOptId(model.Id);
-				
-				if(operation.OperationStatus == 1)
+				// DONE: Başkasının operationunu update edememeliyim.
+				var user = this._userService.GetOnlineUser(this.HttpContext);
+				if (user == null)
+				{
+					return Json(ApiResponse<List<ForumPostModel>>.WithError("Not authorized !"));
+				}
+				var operation = this._operationService.GetById(model.Id);
+				var mission = this._missionService.GetById(operation.MissionId);
+				if(user.Id != mission.UserId)
+				{
+					return Json(ApiResponse<List<ForumPostModel>>.WithError("Not authorized !"));
+				}
+				if(operation.OperationStatus == Enums.OperationStatus.Done)
 				{
 					return Json(ApiResponse<OperationModel>.WithSuccess());
 				}
@@ -181,13 +199,25 @@ namespace TaskApp.Controllers
 				return Json(ApiResponse<OperationModel>.WithError(exp.ToString()));
 			}
 		}
+
 		[HttpDelete]
 		[Route(nameof(DeleteMission))]
 		public ActionResult<ApiResponse> DeleteMission([FromBody] int missionId)
 		{
+			
 			try
 			{
-				
+				// DONE: önce operationsu siliyor sıkıntı yok.
+				var user = this._userService.GetOnlineUser(this.HttpContext);
+				if (user == null)
+				{
+					return Json(ApiResponse<List<ForumPostModel>>.WithError("Not authorized !"));
+				}
+				var mission = this._missionService.GetById(missionId);
+				if(user.Id != mission.UserId)
+				{
+					return Json(ApiResponse<List<ForumPostModel>>.WithError("Not authorized !"));
+				}
 				this._missionService.Delete(missionId);
 
 				return Json(ApiResponse.WithSuccess());
@@ -197,30 +227,31 @@ namespace TaskApp.Controllers
 				return Json(ApiResponse.WithError(exp.ToString()));
 			}
 		}
+
 		[HttpDelete]
 		[Route(nameof(DeleteOperation))]
 		public ActionResult<ApiResponse> DeleteOperation([FromBody] int operationId)
 		{
 			try
 			{
+
+				// DONE: komple toparlanmalı
+				var user = this._userService.GetOnlineUser(this.HttpContext);
 				
-				var operations = this._operationService.GetAll();
-				var OptToBeDeleted = this._operationService.GetByOptId(operationId);
-				List<OperationModel> AllOperations = new List<OperationModel>();
-				int operationCount = 0;
-				foreach (OperationModel operation in operations)
+				if (user == null)
 				{
-					if (operation.MissionId == OptToBeDeleted.MissionId)
-					{
-						operationCount++;
-					}
-				}
-				this._operationService.Delete(operationId);
-				if (operationCount <= 9)
-				{
-					return Json(ApiResponse.WithSuccess());
+					return Json(ApiResponse<List<ForumPostModel>>.WithError("Not authorized !"));
 				}
 
+				var optToBeDeleted = this._operationService.GetById(operationId);
+				var mission = this._missionService.GetById(optToBeDeleted.MissionId);
+				
+				if(user.Id != mission.UserId)
+				{
+					return Json(ApiResponse<List<ForumPostModel>>.WithError("Not authorized !"));
+				}
+
+				this._operationService.Delete(operationId);
 				return Json(ApiResponse.WithSuccess());
 			}
 			catch (Exception exp)
